+++ changelog.md
@@ -7,6 +7,7 @@
 * Fixed bug where is was not possible to create a relation to a table that has a single
   select as primary field.
 * Implemented real time collaboration.
+* Added option to hide fields in a grid view.
 
 ## Released (2021-01-06)